Refactor parser test packages
