+++ compiler/ballerina-parser/src/test/java/io/ballerinalang/compiler/parser/test/syntax/declarations/FunctionDefinitionTest.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.ballerinalang.compiler.parser.test.declarations;
+package io.ballerinalang.compiler.parser.test.syntax.declarations;
 
 import org.testng.annotations.Test;
