+++ compiler/ballerina-parser/src/test/java/io/ballerinalang/compiler/parser/test/syntax/statements/VarDeclarationTest.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.ballerinalang.compiler.parser.test.statements;
+package io.ballerinalang.compiler.parser.test.syntax.statements;
 
 import org.testng.annotations.Test;
